+++ trs/productionsettings.py
@@ -1,0 +1,7 @@
+# He he, no secret production settings here in this open source repository of
+# course :-)
+# We import from a "trs-site" project that's in a private github repo.
+# From here, we import the secret key and so.
+from trs.testsettings import *
+
+from trs_site.productionsettings import *